+++ src/index.js
@@ -224,7 +224,7 @@ export const rules = {
   },
 };
 
-const schemaCache = {};
+let schemaCache = {};
 
 function parseOptions(optionGroup, context) {
   const cacheHit = schemaCache[JSON.stringify(optionGroup)];
@@ -332,6 +332,10 @@ function initSchemaFromString(source) {
 
 
 
+export function clearSchemaCache() {
+  schemaCache = {};
+}
+
 const gqlProcessor = {
   preprocess: function(text) {
     // Wrap the text in backticks and prepend the internal tag. First the text
@@ -360,5 +364,6 @@ export const processors = reduce(gqlFiles, (result, value) => {
 
 export default {
   rules,
-  processors
+  processors,
+  clearSchemaCache
 }